rag_final/RAG_task/demo/requirements.txt
PyPDF2
python-docx
langchain-core
langchain
langchain-community
python-dotenv
openai
tiktoken
azure-search-documents
azure-core
pypdf